+++ packages/google_mobile_ads/lib/src/mobile_ads.dart
@@ -32,7 +32,7 @@ enum AdapterInitializationState {
 class MobileAds {
   MobileAds._();
 
-  static final MobileAds _instance = MobileAds._();
+  static final MobileAds _instance = MobileAds._().._init();
 
   /// Shared instance to initialize the AdMob SDK.
   static MobileAds get instance => _instance;
@@ -44,16 +44,22 @@ class MobileAds {
   ///
   /// If this method is not called, the first ad request automatically
   /// initializes the Google Mobile Ads SDK.
-  Future<InitializationStatus> initialize() {
-    return instanceManager.channel.invokeMethod<InitializationStatus>(
+  Future<InitializationStatus> initialize() async {
+    return (await instanceManager.channel.invokeMethod<InitializationStatus>(
       'MobileAds#initialize',
-    );
+    ))!;
   }
 
   /// Update the [RequestConfiguration] to apply for future ad requests.
   Future<void> updateRequestConfiguration(
       RequestConfiguration requestConfiguration) {
     return instanceManager.updateRequestConfiguration(requestConfiguration);
+  }
+
+  /// Internal init to cleanup state for hot restart.
+  /// This is a workaround for https://github.com/flutter/flutter/issues/7160.
+  void _init() {
+    instanceManager.channel.invokeMethod('_init');
   }
 }