+++ packages/google_mobile_ads/CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.12.1+1
+
+* Fixes a [crash with Swift based native ads](https://github.com/googleads/googleads-mobile-flutter/issues/121)
+
+## 0.12.1
+
+* Rewarded ads now take an optional `ServerSideVerification`, to support [custom data in rewarded ads](https://developers.google.com/admob/ios/rewarded-video-ssv#custom_data).
+
+## 0.12.0
+
+* Migrated to null safety. Minimum Dart SDK version is bumped to 2.12.0.
+
+## 0.11.0+4
+
+* Fixes a [bug](https://github.com/googleads/googleads-mobile-flutter/issues/47) where state is not properly cleaned up on hot restart.
+* Update README and example app to appropriately dispose ads.
+
 ## 0.11.0+3
 
 * Fixes an [Android crash](https://github.com/googleads/googleads-mobile-flutter/issues/46) when reusing Native and Banner Ad objects.